ref(onboarding): Split go onboarding docs
