+++ static/app/gettingStartedDocs/go/echo/onboarding.spec.tsx
@@ -4,7 +4,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './echo';
+import docs from '.';
 
 describe('echo onboarding docs', () => {
   it('renders errors onboarding docs correctly', () => {
