+++ static/app/gettingStartedDocs/go/fasthttp/onboarding.spec.tsx
@@ -4,9 +4,9 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './negroni';
+import docs from '.';
 
-describe('negroni onboarding docs', () => {
+describe('fasthttp onboarding docs', () => {
   it('renders errors onboarding docs correctly', () => {
     renderWithOnboardingLayout(docs);
