+++ static/app/gettingStartedDocs/go/fiber/onboarding.spec.tsx
@@ -4,9 +4,9 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './fasthttp';
+import docs from '.';
 
-describe('fasthttp onboarding docs', () => {
+describe('fiber onboarding docs', () => {
   it('renders errors onboarding docs correctly', () => {
     renderWithOnboardingLayout(docs);
