+++ static/app/gettingStartedDocs/go/go/logs.tsx
@@ -6,9 +6,7 @@ import {
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {t, tct} from 'sentry/locale';
 
-export const getGoLogsOnboarding = <
-  PlatformOptions extends BasePlatformOptions = BasePlatformOptions,
->({
+export const logs = <PlatformOptions extends BasePlatformOptions = BasePlatformOptions>({
   docsPlatform,
 }: {
   docsPlatform: string;
