+++ static/app/gettingStartedDocs/go/http/onboarding.tsx
@@ -1,32 +1,18 @@
 import {ExternalLink} from 'sentry/components/core/link';
 import type {
-  Docs,
   DocsParams,
   OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  getCrashReportGenericInstallSteps,
-  getCrashReportModalConfigDescription,
-  getCrashReportModalIntroduction,
-} from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
-import {
-  feedbackOnboardingJsLoader,
-  replayOnboardingJsLoader,
-} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
 import {t, tct} from 'sentry/locale';
-import {getGoLogsOnboarding} from 'sentry/utils/gettingStartedDocs/go';
 
-type Params = DocsParams;
-
-const getConfigureSnippet = (params: Params) => `
+const getConfigureSnippet = (params: DocsParams) => `
 import (
   "fmt"
   "net/http"
 
   "github.com/getsentry/sentry-go"
-  sentrygin "github.com/getsentry/sentry-go/gin"
-  "github.com/gin-gonic/gin"
+  sentryhttp "github.com/getsentry/sentry-go/http"
 )
 
 // To initialize Sentry's handler, you need to initialize Sentry itself beforehand
@@ -34,7 +20,6 @@ if err := sentry.Init(sentry.ClientOptions{
   Dsn: "${params.dsn.public}",${
     params.isPerformanceSelected
       ? `
-  EnableTracing: true,
   // Set TracesSampleRate to 1.0 to capture 100%
   // of transactions for tracing.
   // We recommend adjusting this value in production,
@@ -51,62 +36,72 @@ if err := sentry.Init(sentry.ClientOptions{
   fmt.Printf("Sentry initialization failed: %v\\n", err)
 }
 
-// Then create your app
-app := gin.Default()
+// Create an instance of sentryhttp
+sentryHandler := sentryhttp.New(sentryhttp.Options{})
 
-// Once it's done, you can attach the handler as one of your middleware
-app.Use(sentrygin.New(sentrygin.Options{}))
+// Once it's done, you can set up routes and attach the handler as one of your middleware
+http.Handle("/", sentryHandler.Handle(&handler{}))
+http.HandleFunc("/foo", sentryHandler.HandleFunc(func(rw http.ResponseWriter, r *http.Request) {
+  panic("y tho")
+}))
 
-// Set up routes
-app.GET("/", func(ctx *gin.Context) {
-  ctx.String(http.StatusOK, "Hello world!")
-})
+fmt.Println("Listening and serving HTTP on :3000")
 
 // And run it
-app.Run(":3000")`;
+if err := http.ListenAndServe(":3000", nil); err != nil {
+  panic(err)
+}`;
 
 const getOptionsSnippet = () => `
 // Whether Sentry should repanic after recovery, in most cases it should be set to true,
-// as gin.Default includes its own Recovery middleware that handles http responses.
+// and you should gracefully handle http responses.
 Repanic bool
 // Whether you want to block the request before moving forward with the response.
-// Because Gin's default "Recovery" handler doesn't restart the application,
-// it's safe to either skip this option or set it to "false".
+// Useful, when you want to restart the process after it panics.
 WaitForDelivery bool
 // Timeout for the event delivery requests.
 Timeout time.Duration`;
 
 const getUsageSnippet = () => `
-app := gin.Default()
+type handler struct{}
 
-app.Use(sentrygin.New(sentrygin.Options{
-  Repanic: true,
-}))
-
-app.Use(func(ctx *gin.Context) {
-  if hub := sentrygin.GetHubFromContext(ctx); hub != nil {
-    hub.Scope().SetTag("someRandomTag", "maybeYouNeedIt")
-  }
-  ctx.Next()
-})
-
-app.GET("/", func(ctx *gin.Context) {
-  if hub := sentrygin.GetHubFromContext(ctx); hub != nil {
+func (h *handler) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
+  if hub := sentry.GetHubFromContext(r.Context()); hub != nil {
     hub.WithScope(func(scope *sentry.Scope) {
       scope.SetExtra("unwantedQuery", "someQueryDataMaybe")
       hub.CaptureMessage("User provided unwanted query string, but we recovered just fine")
     })
   }
-  ctx.Status(http.StatusOK)
+  rw.WriteHeader(http.StatusOK)
+}
+
+func enhanceSentryEvent(handler http.HandlerFunc) http.HandlerFunc {
+  return func(rw http.ResponseWriter, r *http.Request) {
+    if hub := sentry.GetHubFromContext(r.Context()); hub != nil {
+      hub.Scope().SetTag("someRandomTag", "maybeYouNeedIt")
+    }
+    handler(rw, r)
+  }
+}
+
+// Later in the code
+
+sentryHandler := sentryhttp.New(sentryhttp.Options{
+  Repanic: true,
 })
 
-app.GET("/foo", func(ctx *gin.Context) {
-  // sentrygin handler will catch it just fine. Also, because we attached "someRandomTag"
-  // in the middleware before, it will be sent through as well
-  panic("y tho")
-})
+http.Handle("/", sentryHandler.Handle(&handler{}))
+http.HandleFunc("/foo", sentryHandler.HandleFunc(
+  enhanceSentryEvent(func(rw http.ResponseWriter, r *http.Request) {
+    panic("y tho")
+  }),
+))
 
-app.Run(":3000")`;
+fmt.Println("Listening and serving HTTP on :3000")
+
+if err := http.ListenAndServe(":3000", nil); err != nil {
+  panic(err)
+}`;
 
 const getBeforeSendSnippet = (params: any) => `
 sentry.Init(sentry.ClientOptions{
@@ -122,21 +117,21 @@ sentry.Init(sentry.ClientOptions{
   },
 })`;
 
-const onboarding: OnboardingConfig = {
+export const onboarding: OnboardingConfig = {
   install: () => [
     {
       type: StepType.INSTALL,
       content: [
         {
           type: 'text',
-          text: tct('Install our Go Gin SDK using [code:go get]:', {
+          text: tct('Install our Go HTTP SDK using [code:go get]:', {
             code: <code />,
           }),
         },
         {
           type: 'code',
           language: 'bash',
-          code: 'go get github.com/getsentry/sentry-go/gin',
+          code: 'go get github.com/getsentry/sentry-go/http',
         },
       ],
     },
@@ -164,7 +159,7 @@ const onboarding: OnboardingConfig = {
           type: 'text',
           text: [
             tct(
-              '[code:sentrygin] accepts a struct of [code:Options] that allows you to configure how the handler will behave.',
+              '[code:sentryhttp] accepts a struct of [code:Options] that allows you to configure how the handler will behave.',
               {code: <code />}
             ),
             t('Currently it respects 3 options:'),
@@ -183,7 +178,7 @@ const onboarding: OnboardingConfig = {
         {
           type: 'text',
           text: tct(
-            "[code:sentrygin] attaches an instance of [sentryHubLink:*sentry.Hub] to the [code:*gin.Context], which makes it available throughout the rest of the request's lifetime. You can access it by using the [code:sentrygin.GetHubFromContext()] method on the context itself in any of your proceeding middleware and routes. And it should be used instead of the global [code:sentry.CaptureMessage], [code:sentry.CaptureException], or any other calls, as it keeps the separation of data between the requests.",
+            "[code:sentryhttp] attaches an instance of [sentryHubLink:*sentry.Hub] to the request's context, which makes it available throughout the rest of the request's lifetime. You can access it by using the [code:sentry.GetHubFromContext()] method on the request itself in any of your proceeding middleware and routes. And it should be used instead of the global [code:sentry.CaptureMessage], [code:sentry.CaptureException], or any other calls, as it keeps the separation of data between the requests.",
             {
               code: <code />,
               sentryHubLink: (
@@ -197,7 +192,7 @@ const onboarding: OnboardingConfig = {
           alertType: 'info',
           showIcon: false,
           text: tct(
-            "Keep in mind that [code:*sentry.Hub] won't be available in middleware attached before [code:sentrygin]!",
+            "Keep in mind that [code:*sentry.Hub] won't be available in middleware attached before [code:sentryhttp]!",
             {code: <code />}
           ),
         },
@@ -221,7 +216,7 @@ const onboarding: OnboardingConfig = {
     },
   ],
   verify: () => [],
-  nextSteps: (params: Params) => {
+  nextSteps: (params: DocsParams) => {
     const steps = [];
 
     if (params.isLogsSelected) {
@@ -238,35 +233,3 @@ const onboarding: OnboardingConfig = {
     return steps;
   },
 };
-
-const crashReportOnboarding: OnboardingConfig = {
-  introduction: () => getCrashReportModalIntroduction(),
-  install: (params: Params) => getCrashReportGenericInstallSteps(params),
-  configure: () => [
-    {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: getCrashReportModalConfigDescription({
-            link: 'https://docs.sentry.io/platforms/go/guides/gin/user-feedback/configuration/#crash-report-modal',
-          }),
-        },
-      ],
-    },
-  ],
-  verify: () => [],
-  nextSteps: () => [],
-};
-
-const docs: Docs = {
-  onboarding,
-  replayOnboardingJsLoader,
-  crashReportOnboarding,
-  feedbackOnboardingJsLoader,
-  logsOnboarding: getGoLogsOnboarding({
-    docsPlatform: 'gin',
-  }),
-};
-
-export default docs;
