+++ static/app/gettingStartedDocs/go/iris/onboarding.tsx
@@ -1,31 +1,18 @@
 import {ExternalLink} from 'sentry/components/core/link';
 import type {
-  Docs,
   DocsParams,
   OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  getCrashReportGenericInstallSteps,
-  getCrashReportModalConfigDescription,
-  getCrashReportModalIntroduction,
-} from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
-import {
-  feedbackOnboardingJsLoader,
-  replayOnboardingJsLoader,
-} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
 import {t, tct} from 'sentry/locale';
-import {getGoLogsOnboarding} from 'sentry/utils/gettingStartedDocs/go';
 
-type Params = DocsParams;
-
-const getConfigureSnippet = (params: Params) => `
+const getConfigureSnippet = (params: DocsParams) => `
 import (
   "fmt"
-  "net/http"
 
   "github.com/getsentry/sentry-go"
-  sentryhttp "github.com/getsentry/sentry-go/http"
+  sentryiris "github.com/getsentry/sentry-go/iris"
+  "github.com/kataras/iris/v12"
 )
 
 // To initialize Sentry's handler, you need to initialize Sentry itself beforehand
@@ -49,72 +36,61 @@ if err := sentry.Init(sentry.ClientOptions{
   fmt.Printf("Sentry initialization failed: %v\\n", err)
 }
 
-// Create an instance of sentryhttp
-sentryHandler := sentryhttp.New(sentryhttp.Options{})
+// Then create your app
+app := iris.Default()
 
-// Once it's done, you can set up routes and attach the handler as one of your middleware
-http.Handle("/", sentryHandler.Handle(&handler{}))
-http.HandleFunc("/foo", sentryHandler.HandleFunc(func(rw http.ResponseWriter, r *http.Request) {
-  panic("y tho")
-}))
+// Once it's done, you can attach the handler as one of your middleware
+app.Use(sentryiris.New(sentryiris.Options{}))
 
-fmt.Println("Listening and serving HTTP on :3000")
+// Set up routes
+app.Get("/", func(ctx iris.Context) {
+  ctx.Writef("Hello world!")
+})
 
 // And run it
-if err := http.ListenAndServe(":3000", nil); err != nil {
-  panic(err)
-}`;
+app.Run(iris.Addr(":3000"))`;
 
 const getOptionsSnippet = () => `
 // Whether Sentry should repanic after recovery, in most cases it should be set to true,
-// and you should gracefully handle http responses.
+// as iris.Default includes its own Recovery middleware what handles http responses.
 Repanic bool
 // Whether you want to block the request before moving forward with the response.
-// Useful, when you want to restart the process after it panics.
+// Because Iris's default "Recovery" handler doesn't restart the application,
+// it's safe to either skip this option or set it to "false".
 WaitForDelivery bool
 // Timeout for the event delivery requests.
 Timeout time.Duration`;
 
 const getUsageSnippet = () => `
-type handler struct{}
+app := iris.Default()
 
-func (h *handler) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
-  if hub := sentry.GetHubFromContext(r.Context()); hub != nil {
+app.Use(sentryiris.New(sentryiris.Options{
+  Repanic: true,
+}))
+
+app.Use(func(ctx iris.Context) {
+  if hub := sentryiris.GetHubFromContext(ctx); hub != nil {
+    hub.Scope().SetTag("someRandomTag", "maybeYouNeedIt")
+  }
+  ctx.Next()
+})
+
+app.Get("/", func(ctx iris.Context) {
+  if hub := sentryiris.GetHubFromContext(ctx); hub != nil {
     hub.WithScope(func(scope *sentry.Scope) {
       scope.SetExtra("unwantedQuery", "someQueryDataMaybe")
       hub.CaptureMessage("User provided unwanted query string, but we recovered just fine")
     })
   }
-  rw.WriteHeader(http.StatusOK)
-}
-
-func enhanceSentryEvent(handler http.HandlerFunc) http.HandlerFunc {
-  return func(rw http.ResponseWriter, r *http.Request) {
-    if hub := sentry.GetHubFromContext(r.Context()); hub != nil {
-      hub.Scope().SetTag("someRandomTag", "maybeYouNeedIt")
-    }
-    handler(rw, r)
-  }
-}
-
-// Later in the code
-
-sentryHandler := sentryhttp.New(sentryhttp.Options{
-  Repanic: true,
 })
 
-http.Handle("/", sentryHandler.Handle(&handler{}))
-http.HandleFunc("/foo", sentryHandler.HandleFunc(
-  enhanceSentryEvent(func(rw http.ResponseWriter, r *http.Request) {
-    panic("y tho")
-  }),
-))
+app.Get("/foo", func(ctx iris.Context) {
+  // sentryiris handler will catch it just fine. Also, because we attached "someRandomTag"
+  // in the middleware before, it will be sent through as well
+  panic("y tho")
+})
 
-fmt.Println("Listening and serving HTTP on :3000")
-
-if err := http.ListenAndServe(":3000", nil); err != nil {
-  panic(err)
-}`;
+app.Run(iris.Addr(":3000"))`;
 
 const getBeforeSendSnippet = (params: any) => `
 sentry.Init(sentry.ClientOptions{
@@ -130,21 +106,21 @@ sentry.Init(sentry.ClientOptions{
   },
 })`;
 
-const onboarding: OnboardingConfig = {
+export const onboarding: OnboardingConfig = {
   install: () => [
     {
       type: StepType.INSTALL,
       content: [
         {
           type: 'text',
-          text: tct('Install our Go HTTP SDK using [code:go get]:', {
+          text: tct('Install our Go Iris SDK using [code:go get]:', {
             code: <code />,
           }),
         },
         {
           type: 'code',
           language: 'bash',
-          code: 'go get github.com/getsentry/sentry-go/http',
+          code: 'go get github.com/getsentry/sentry-go/iris',
         },
       ],
     },
@@ -172,7 +148,7 @@ const onboarding: OnboardingConfig = {
           type: 'text',
           text: [
             tct(
-              '[code:sentryhttp] accepts a struct of [code:Options] that allows you to configure how the handler will behave.',
+              '[code:sentryiris] accepts a struct of [code:Options] that allows you to configure how the handler will behave.',
               {code: <code />}
             ),
             t('Currently it respects 3 options:'),
@@ -191,7 +167,7 @@ const onboarding: OnboardingConfig = {
         {
           type: 'text',
           text: tct(
-            "[code:sentryhttp] attaches an instance of [sentryHubLink:*sentry.Hub] to the request's context, which makes it available throughout the rest of the request's lifetime. You can access it by using the [code:sentry.GetHubFromContext()] method on the request itself in any of your proceeding middleware and routes. And it should be used instead of the global [code:sentry.CaptureMessage], [code:sentry.CaptureException], or any other calls, as it keeps the separation of data between the requests.",
+            "[code:sentryiris] attaches an instance of [sentryHubLink:*sentry.Hub] to the [code:iris.Context], which makes it available throughout the rest of the request's lifetime. You can access it by using the [code:sentryiris.GetHubFromContext()] method on the context itself in any of your proceeding middleware and routes. And it should be used instead of the global [code:sentry.CaptureMessage], [code:sentry.CaptureException], or any other calls, as it keeps the separation of data between the requests.",
             {
               code: <code />,
               sentryHubLink: (
@@ -205,7 +181,7 @@ const onboarding: OnboardingConfig = {
           alertType: 'info',
           showIcon: false,
           text: tct(
-            "Keep in mind that [code:*sentry.Hub] won't be available in middleware attached before [code:sentryhttp]!",
+            "Keep in mind that [code:*sentry.Hub] won't be available in middleware attached before [code:sentryiris]!",
             {code: <code />}
           ),
         },
@@ -229,7 +205,7 @@ const onboarding: OnboardingConfig = {
     },
   ],
   verify: () => [],
-  nextSteps: (params: Params) => {
+  nextSteps: (params: DocsParams) => {
     const steps = [];
 
     if (params.isLogsSelected) {
@@ -246,35 +222,3 @@ const onboarding: OnboardingConfig = {
     return steps;
   },
 };
-
-const crashReportOnboarding: OnboardingConfig = {
-  introduction: () => getCrashReportModalIntroduction(),
-  install: (params: Params) => getCrashReportGenericInstallSteps(params),
-  configure: () => [
-    {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: getCrashReportModalConfigDescription({
-            link: 'https://docs.sentry.io/platforms/go/guides/http/user-feedback/configuration/#crash-report-modal',
-          }),
-        },
-      ],
-    },
-  ],
-  verify: () => [],
-  nextSteps: () => [],
-};
-
-const docs: Docs = {
-  onboarding,
-  replayOnboardingJsLoader,
-  crashReportOnboarding,
-  feedbackOnboardingJsLoader,
-  logsOnboarding: getGoLogsOnboarding({
-    docsPlatform: 'http',
-  }),
-};
-
-export default docs;
