+++ static/app/gettingStartedDocs/go/martini/onboarding.spec.tsx
@@ -4,9 +4,9 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './gin';
+import docs from '.';
 
-describe('gin onboarding docs', () => {
+describe('martini onboarding docs', () => {
   it('renders errors onboarding docs correctly', () => {
     renderWithOnboardingLayout(docs);
