+++ static/app/gettingStartedDocs/go/martini/onboarding.tsx
@@ -1,6 +1,5 @@
 import {ExternalLink} from 'sentry/components/core/link';
 import type {
-  Docs,
   DocsParams,
   OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
@@ -10,23 +9,15 @@ import {
   getCrashReportModalConfigDescription,
   getCrashReportModalIntroduction,
 } from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
-import {
-  feedbackOnboardingJsLoader,
-  replayOnboardingJsLoader,
-} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
 import {t, tct} from 'sentry/locale';
-import {getGoLogsOnboarding} from 'sentry/utils/gettingStartedDocs/go';
 
-type Params = DocsParams;
-
-const getConfigureSnippet = (params: Params) => `
+const getConfigureSnippet = (params: DocsParams) => `
 import (
   "fmt"
-  "net/http"
 
   "github.com/getsentry/sentry-go"
-  sentrynegroni "github.com/getsentry/sentry-go/negroni"
-  "github.com/urfave/negroni"
+  sentrymartini "github.com/getsentry/sentry-go/martini"
+  "github.com/go-martini/martini"
 )
 
 // To initialize Sentry's handler, you need to initialize Sentry itself beforehand
@@ -51,67 +42,58 @@ if err := sentry.Init(sentry.ClientOptions{
 }
 
 // Then create your app
-app := negroni.Classic()
+app := martini.Classic()
 
 // Once it's done, you can attach the handler as one of your middleware
-app.Use(sentrynegroni.New(sentrynegroni.Options{}))
+app.Use(sentrymartini.New(sentrymartini.Options{}))
 
 // Set up routes
-mux := http.NewServeMux()
-
-mux.HandleFunc("/", func(rw http.ResponseWriter, r *http.Request) {
-  fmt.Fprintf(w, "Hello world!")
+app.Get("/", func() string {
+  return "Hello world!"
 })
 
-app.UseHandler(mux)
-
 // And run it
-http.ListenAndServe(":3000", app)`;
+app.Run()`;
 
 const getOptionsSnippet = () => `
 // Whether Sentry should repanic after recovery, in most cases it should be set to true,
-// as negroni.Classic includes its own Recovery middleware that handles http responses.
+// as martini.Classic includes its own Recovery middleware that handles http responses.
 Repanic bool
 // Whether you want to block the request before moving forward with the response.
-// Because Negroni's default "Recovery" handler doesn't restart the application,
+// Because Martini's default "Recovery" handler doesn't restart the application,
 // it's safe to either skip this option or set it to "false".
 WaitForDelivery bool
 // Timeout for the event delivery requests.
 Timeout time.Duration`;
 
 const getUsageSnippet = () => `
-app := negroni.Classic()
+app := martini.Classic()
 
-app.Use(sentrynegroni.New(sentrynegroni.Options{
+app.Use(sentrymartini.New(sentrymartini.Options{
   Repanic: true,
 }))
 
-app.Use(negroni.HandlerFunc(func(rw http.ResponseWriter, r *http.Request, next http.HandlerFunc) {
-  hub := sentry.GetHubFromContext(r.Context())
+app.Use(func(rw http.ResponseWriter, r *http.Request, c martini.Context, hub *sentry.Hub) {
   hub.Scope().SetTag("someRandomTag", "maybeYouNeedIt")
-  next(rw, r)
-}))
+})
 
-mux := http.NewServeMux()
-
-mux.HandleFunc("/", func(rw http.ResponseWriter, r *http.Request) {
-  hub := sentry.GetHubFromContext(r.Context())
-  hub.WithScope(func(scope *sentry.Scope) {
-    scope.SetExtra("unwantedQuery", "someQueryDataMaybe")
-    hub.CaptureMessage("User provided unwanted query string, but we recovered just fine")
-  })
+app.Get("/", func(rw http.ResponseWriter, r *http.Request, hub *sentry.Hub) {
+  if someCondition {
+    hub.WithScope(func (scope *sentry.Scope) {
+      scope.SetExtra("unwantedQuery", rw.URL.RawQuery)
+      hub.CaptureMessage("User provided unwanted query string, but we recovered just fine")
+    })
+  }
   rw.WriteHeader(http.StatusOK)
 })
 
-mux.HandleFunc("/foo", func(rw http.ResponseWriter, r *http.Request) {
-  // sentrynagroni handler will catch it just fine. Also, because we attached "someRandomTag"
+app.Get("/foo", func() string {
+  // sentrymartini handler will catch it just fine. Also, because we attached "someRandomTag"
   // in the middleware before, it will be sent through as well
   panic("y tho")
 })
 
-app.UseHandler(mux)
-
-http.ListenAndServe(":3000", app)`;
+app.Run()`;
 
 const getBeforeSendSnippet = (params: any) => `
 sentry.Init(sentry.ClientOptions{
@@ -127,38 +109,21 @@ sentry.Init(sentry.ClientOptions{
   },
 })`;
 
-const getPanicHandlerSnippet = () => `
-app := negroni.New()
-
-recovery := negroni.NewRecovery()
-recovery.PanicHandlerFunc = sentrynegroni.PanicHandlerFunc
-
-app.Use(recovery)
-
-mux := http.NewServeMux()
-mux.HandleFunc("/", func(w http.ResponseWriter, req *http.Request) {
-  panic("y tho")
-})
-
-app.UseHandler(mux)
-
-http.ListenAndServe(":3000", app)`;
-
-const onboarding: OnboardingConfig = {
+export const onboarding: OnboardingConfig = {
   install: () => [
     {
       type: StepType.INSTALL,
       content: [
         {
           type: 'text',
-          text: tct('Install our Go Negroni SDK using [code:go get]:', {
+          text: tct('Install our Go Martini SDK using [code:go get]:', {
             code: <code />,
           }),
         },
         {
           type: 'code',
           language: 'bash',
-          code: 'go get github.com/getsentry/sentry-go/negroni',
+          code: 'go get github.com/getsentry/sentry-go/martini',
         },
       ],
     },
@@ -186,7 +151,7 @@ const onboarding: OnboardingConfig = {
           type: 'text',
           text: [
             tct(
-              '[code:sentrynegroni] accepts a struct of [code:Options] that allows you to configure how the handler will behave.',
+              '[code:sentrymartini] accepts a struct of [code:Options] that allows you to configure how the handler will behave.',
               {code: <code />}
             ),
             t('Currently it respects 3 options:'),
@@ -205,7 +170,7 @@ const onboarding: OnboardingConfig = {
         {
           type: 'text',
           text: tct(
-            "[code:sentrynegroni] attaches an instance of [sentryHubLink:*sentry.Hub] to the request's context, which makes it available throughout the rest of the request's lifetime. You can access it by using the [code:sentry.GetHubFromContext()] method on the request itself in any of your proceeding middleware and routes. And it should be used instead of the global [code:sentry.CaptureMessage], [code:sentry.CaptureException], or any other calls, as it keeps the separation of data between the requests.",
+            "[code:sentrymartini] maps an instance of [sentryHubLink:*sentry.Hub] as one of the services available throughout the rest of the request's lifetime. You can access it by providing a hub [code:*sentry.Hub] parameter in any of your proceeding middleware and routes. And it should be used instead of the global [code:sentry.CaptureMessage], [code:sentry.CaptureException], or any other calls, as it keeps the separation of data between the requests.",
             {
               code: <code />,
               sentryHubLink: (
@@ -219,7 +184,7 @@ const onboarding: OnboardingConfig = {
           alertType: 'info',
           showIcon: false,
           text: tct(
-            "Keep in mind that [code:*sentry.Hub] won't be available in middleware attached before [code:sentrynegroni]!",
+            "Keep in mind that [code:*sentry.Hub] won't be available in middleware attached before [code:sentrymartini]!",
             {code: <code />}
           ),
         },
@@ -241,46 +206,9 @@ const onboarding: OnboardingConfig = {
         },
       ],
     },
-    {
-      title: t("Using Negroni's 'panicHandlerFuncCode' Option"),
-      content: [
-        {
-          type: 'text',
-          text: tct(
-            "Negroni provides an option called [code:PanicHandlerFunc], which lets you 'plug-in' to its default [code:Recovery] middleware.",
-            {
-              code: <code />,
-            }
-          ),
-        },
-        {
-          type: 'text',
-          text: tct(
-            "[sentrynegroniCode:sentrynegroni] exports a very barebones implementation, which utilizes it, so if you don't need anything else than just reporting panics to Sentry, you can use it instead, as it's just one line of code!",
-            {
-              sentrynegroniCode: <code />,
-            }
-          ),
-        },
-        {
-          type: 'text',
-          text: tct(
-            'You can still use [beforeSendCode:BeforeSend] and event processors to modify data before delivering it to Sentry, using this method as well.',
-            {
-              beforeSendCode: <code />,
-            }
-          ),
-        },
-        {
-          type: 'code',
-          language: 'go',
-          code: getPanicHandlerSnippet(),
-        },
-      ],
-    },
   ],
   verify: () => [],
-  nextSteps: (params: Params) => {
+  nextSteps: (params: DocsParams) => {
     const steps = [];
 
     if (params.isLogsSelected) {
@@ -298,9 +226,9 @@ const onboarding: OnboardingConfig = {
   },
 };
 
-const crashReportOnboarding: OnboardingConfig = {
+export const crashReport: OnboardingConfig = {
   introduction: () => getCrashReportModalIntroduction(),
-  install: (params: Params) => getCrashReportGenericInstallSteps(params),
+  install: (params: DocsParams) => getCrashReportGenericInstallSteps(params),
   configure: () => [
     {
       type: StepType.CONFIGURE,
@@ -308,7 +236,7 @@ const crashReportOnboarding: OnboardingConfig = {
         {
           type: 'text',
           text: getCrashReportModalConfigDescription({
-            link: 'https://docs.sentry.io/platforms/go/guides/negroni/user-feedback/configuration/#crash-report-modal',
+            link: 'https://docs.sentry.io/platforms/go/guides/martini/user-feedback/configuration/#crash-report-modal',
           }),
         },
       ],
@@ -317,15 +245,3 @@ const crashReportOnboarding: OnboardingConfig = {
   verify: () => [],
   nextSteps: () => [],
 };
-
-const docs: Docs = {
-  onboarding,
-  replayOnboardingJsLoader,
-  crashReportOnboarding,
-  feedbackOnboardingJsLoader,
-  logsOnboarding: getGoLogsOnboarding({
-    docsPlatform: 'negroni',
-  }),
-};
-
-export default docs;
