+++ static/app/gettingStartedDocs/go/negroni/index.tsx
@@ -1,0 +1,23 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {crashReport} from 'sentry/gettingStartedDocs/go/go/crashReport';
+import {logs} from 'sentry/gettingStartedDocs/go/go/logs';
+import {onboarding} from 'sentry/gettingStartedDocs/go/negroni/onboarding';
+import {
+  feedbackOnboardingJsLoader,
+  replayOnboardingJsLoader,
+} from 'sentry/gettingStartedDocs/javascript/jsLoader/jsLoader';
+
+const docs: Docs = {
+  onboarding,
+  replayOnboardingJsLoader,
+  crashReportOnboarding: crashReport({
+    docsLink:
+      'https://docs.sentry.io/platforms/go/guides/negroni/user-feedback/configuration/#crash-report-modal',
+  }),
+  feedbackOnboardingJsLoader,
+  logsOnboarding: logs({
+    docsPlatform: 'negroni',
+  }),
+};
+
+export default docs;
