+++ static/app/gettingStartedDocs/go/negroni/onboarding.spec.tsx
@@ -4,9 +4,9 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './fiber';
+import docs from '.';
 
-describe('fiber onboarding docs', () => {
+describe('negroni onboarding docs', () => {
   it('renders errors onboarding docs correctly', () => {
     renderWithOnboardingLayout(docs);
